setDate() needs to be called (this addresses #3)
Need to explicitly setDate in order to properly propogate all of a
story's data into a page. This hack checks to see if setDate() has EVER
been run.

In the initial case of being sent through Meteor.Router from landing ->
journal, this is not the case. So upon page render, run setDate() if
never run.

Does this raise more issues? I need to be sure that ANYTIME we want to
transition to a new date, setDate() has to be called. Need a more
centralized way of ensuring this, rather than random calls from all
through out the code.